distribFunFromPDF_2D_1D, camera size fixed, tested blur on simulation
woah! It's working! Also, 4e-6 is a pixel size, not the camera size, so I was always looking at stuff 100 times smaller (hopefully, I still have to confirm)

+++ testDiffraction_thirdAttempt.py
@@ -142,7 +142,7 @@ class mappedFunction:
 		self.f=f
 		self.center=center.astype(float)
 		self.size=size.astype(float)
-		self.resolution=resolution
+		self.resolution=resolution.astype(int)
 		self.subRange=subRange
